+++ Code/Content/Actions.cs
@@ -17,6 +17,12 @@ internal static class WarBoxActions
         WarBox.LogInfo("Army Limits toggled to " + pCurrentValue);
     }
 
+    public static void UpdateArmyPerPop(int pCurrentValue)
+    {
+        WarBox.warbox_army_per_pop = pCurrentValue;
+        WarBox.LogInfo("Army Limits toggled to " + pCurrentValue);
+    }
+
     public static bool LaunchATGM(Actor pSelf, Vector2 pTarget, WorldTile pTile = null)
     {
         Vector2Int pos = pTile.pos;
@@ -26,8 +32,6 @@ internal static class WarBoxActions
         Vector3 newPoint2 = Toolbox.getNewPoint(vector.x, vector.y, pos.x, pos.y, pSelf.a.stats["size"]);
         newPoint2.y += 0.5f;
         World.world.projectiles.spawn(pSelf, null, "rpg_projectile", newPoint2, newPoint);
-
-        //pSelf.punchTargetAnimation(pTarget, pFlip: true, pReverse: false, 45f);
         return true;
     }